django_api/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cffi==1.15.1
cryptography==38.0.4
Django==4.1.3
django-environ==0.9.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
importlib-metadata==5.1.0
Markdown==3.4.1
MarkupSafe==2.1.1
psycopg2-binary==2.9.5
pycparser==2.21
pyOpenSSL==22.1.0
pytz==2022.6
sqlparse==0.4.3
Werkzeug==2.2.2
zipp==3.11.0
